Listar e deletar formas

+++ src/main/java/com/mycompany/myapp/service/FormaService.java
@@ -1,6 +1,7 @@
 package com.mycompany.myapp.service;
 
 import com.mycompany.myapp.domain.Forma;
+import com.mycompany.myapp.domain.Poligono;
 import com.mycompany.myapp.repository.FormaRepository;
 import com.mycompany.myapp.service.dto.FormaDTO;
 import com.mycompany.myapp.service.mapper.FormaMapper;
@@ -105,8 +106,14 @@ public class FormaService {
      *
      * @param id the id of the entity.
      */
+    @Transactional
     public void delete(Long id) {
         log.debug("Request to delete Forma : {}", id);
+        //formaRepository.deleteBy
+        Optional<Forma> forma = formaRepository.findById(id);
+        if (forma.isEmpty()) throw new IllegalArgumentException();
+        for (Forma f : forma.get().getFormas()) f.setAgrupamento(null);
+        for (Poligono p : forma.get().getPoligonos()) p.setForma(null);
         formaRepository.deleteById(id);
     }
 }
